Implemented fixes as requested
"make sure you tag all your functions properly. So assign them client/Server and so on when needed this helps stop hackers, Next we found that you had network ids on almost every object in the scene and most of them don't need it like for instance the walls. Try not to find game objects with strings its not reliable.  Also check you comments try not to leave any WIP stuff in there good luck."